requirements.txt
django
djangorestframework
psycopg[binary]
pillow
django-environ
